适配3.10新core

+++ WangPluginSav/GUI/WonderRecordsNX/WonderRecordSWSH.cs
@@ -1711,7 +1711,7 @@ partial class WonderRecordSWSH : Form
             case 0:
                 return null;
             case 1:
-                image = SpriteUtil.GetSprite(num2, b, num6, 0u, num5, flag, shiny, EntityContext.None);
+                image = SpriteUtil.GetSprite(num2, b, (byte)num6, 0u, num5, flag, shiny, EntityContext.None);
                 if (flag2)
                 {
                     Bitmap dyna = PKHeX.Drawing.PokeSprite.Properties.Resources.dyna;
